move infinitearrays to its own file
